Move colormap to base fractool directory (for now... seems like a better place)
