+++ include/fractool/colormap.hpp
